backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
numpy<2.0.0
sentence-transformers>=2.5.0
pymupdf==1.23.8
newspaper3k==0.2.8
beautifulsoup4==4.12.2
lxml_html_clean>=0.1.0
requests==2.31.0
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0

